requirements.txt
streamlit
pandas
rapidfuzz
openai==0.28
requests
google-generativeai
python-dotenv